// src/model/structure-definition.ts
export interface StructureDefinition extends DomainResource {
  // === Resource Identity ===

  /** Resource type discriminator (1..1) */
  resourceType: 'StructureDefinition';

  // === Conformance Resource Metadata ===

  /**
   * Canonical identifier for this structure definition, represented as a URI
   * (globally unique) (1..1)
   * @see https://hl7.org/fhir/R4/structuredefinition-definitions.html#StructureDefinition.url
   */
  url: FhirUri;

  /**
   * Additional identifier for the structure definition (0..*)
   * @see https://hl7.org/fhir/R4/structuredefinition-definitions.html#StructureDefinition.identifier
   */
  identifier?: Identifier[];

  /**
   * Business version of the structure definition (0..1)
   * @see https://hl7.org/fhir/R4/structuredefinition-definitions.html#StructureDefinition.version
   */
  version?: FhirString;
// ...
  kind: StructureDefinitionKind;
// ...
  type: FhirUri;
// ...
}